Updates to cat page
Changed the way I was displaying data to the front with cards. Using a different file to create my card. I then pass the parameters to were my card function is located. Cards are now aligning in two columns and displaying the data. CatPage.js look cleaner but I still need to finish cleaning the code and styling everything correctly.